move fixtures from conftest to fixtures dir
